benchmark/requirements.txt
rlms>=0.1.1
vllm>=0.8.0
datasets>=3.0.0
matplotlib>=3.9.0
nvidia-ml-py3>=7.352.0
numpy>=1.26.0
